added company to queue test

+++ server/test/services/serviceSync.spec.js
@@ -5,8 +5,10 @@ const mongo = require('../helpers/mongo');
 const zoho = require('../../controller/services/zoho/zoho');
 const docGenerator = require('../helpers/docGenerator-v2');
 const candidateHelper = require('../api-v2/users/candidates/candidateHelpers');
+const companyHelper = require('../api-v2/users/companies/companyHelpers');
 const syncQueue = require('../../model/mongoose/sync_queue');
 const users = require('../../model/mongoose/users');
+const companies = require('../../model/mongoose/companies');
 const serviceSync = require('../../controller/services/serviceSync');
 const sendgrid = require('../../controller/services/email/sendGrid');
 
@@ -18,7 +20,7 @@ chai.use(chaiHttp);
 const testContact = {
     email: "testingemail@workonblockchain.com",
     first_name: "PART OF AUTOMATIC UNIT TESTS",
-    last_name: "WILL DELETE AUTOMATICALLY"
+    company_name: "PART OF AUTOMATIC UNIT TESTS"
 };
 const syncTestEmail = sendgrid.addEmailEnvironment(testContact.email);
 const getSyncTestEmail = syncTestEmail.replace("+","%2B");
@@ -36,9 +38,11 @@ describe('service syncronization', function () {
                 email: getSyncTestEmail
             }
         });
-        await zoho.contacts.deleteOne({
-            id: res[0].id
-        });
+        if (res && res.length > 0) {
+            await zoho.contacts.deleteOne({
+                id: res[0].id
+            });
+        }
         console.log('dropping database');
         await mongo.drop();
     })
@@ -49,7 +53,6 @@ describe('service syncronization', function () {
             const candidate = docGenerator.candidate();
             candidate.email = testContact.email;
             candidate.first_name = testContact.first_name;
-            candidate.last_name = testContact.last_name;
 
             await candidateHelper.signupCandidate(candidate);
 
@@ -70,7 +73,6 @@ describe('service syncronization', function () {
             const candidate = docGenerator.candidate();
             candidate.email = testContact.email;
             candidate.first_name = testContact.first_name;
-            candidate.last_name = testContact.last_name;
             const profileData = docGenerator.candidateProfile();
 
             await candidateHelper.candidateProfile(candidate, profileData);
@@ -95,6 +97,35 @@ describe('service syncronization', function () {
             });
             zohoContact[0].Candidate_status.should.equal(userDoc.candidate.latest_status.status);
             zohoContact[0].Last_Name.should.equal(userDoc.last_name);
+        })
+
+        it('should sync a new company', async function () {
+            const company = docGenerator.company();
+            company.email = testContact.email;
+            company.first_name = testContact.first_name;
+            company.company_name = testContact.company_name;
+
+            await companyHelper.signupCompany(company);
+
+            await serviceSync.pullFromQueue();
+
+            const userDoc = await users.findOneByEmail(company.email);
+            const companyDoc = await companies.findOneByEmail(company.email);
+            const zohoContact = await zoho.contacts.search({
+                params: {
+                    email: getSyncTestEmail
+                }
+            });
+            const zohoAccount = await zoho.accounts.search({
+                params: {
+                    criteria: "((Account_Name:equals:" + company.company_name + "))"
+                }
+            });
+            zohoContact[0].First_Name.should.equal(userDoc.first_name);
+            zohoContact[0].Candidate_status.should.equal(userDoc.candidate.latest_status.status);
+            zohoContact[0].Last_Name.should.equal(userDoc.last_name);
+            zohoAccount[0].Account_Name.should.equal(companyDoc.company_name);
+            zohoAccount[0].Account_Name.should.equal(companyDoc.company_name);
         })
     })
 })
